ambientsync: simplify and support all corridor formats
Instead of adding a dummy entry at the end of the wind corridor we now
save the original corridor and replace it with a single entry.

This means we can now support all corridor formats without having to
learn how to identify the format used and generate a matching dummy
waypoint.

The only downside with this approach is that we will not have the
correct descent winds available as long as we're above the TL (below
the TL the original corridor is restored).

Either accept this (PSX will interpolate decent descent winds from the
active wind when you load the descent winds into the FMC) or briefly
switch the altimeter to QNH/1013, load the descent winds and then
switch back to STD.