Add frankenfreeze.py
This is an attempt to improve PSX icing by using the MSFS in-cloud
status (from SimConnect) to force PSX to be in a cloud layer (or not
be in a cloud layer).

If the PSX TAT (which will usually match MSFS TAT well) is in the
icing range, icing will happen when you enter MSFS clouds.

The script tries to modify PSX's weather as little as possible to
achieve this.

This will only affect PSX icing caused by being inside a cloud
layer. Icing caused by fog or precipitation is still handled
completely by PSX, but since those occur close to the ground, i.e
close to an airport that has a METAR, the PSX and MSFS weather should
not differ much.