frankenfreeze: updated to be able to adjust PSX OAT
I added this to allow tweaking of PSX's OAT and not just the cloud
layers in order to get even closer to MSFS icing conditions.

But based on the logging I did during some Worldflight flights, it
seems this is not needed, it will not make things much better, and it
likely to affect other things in PSX. But I'm leaving the code in, but
disabled by default.