frankenusb: implement runway entry/exit/etc. macros
A kind of lightweight virtual copilot. Very VKB
STECS-focused (triggered by pressing one button and the aircraft
config then changes based on the position of a rotary 6-position
switch).

If anyone else wants to do something similar we can restruture the
code to be triggered in more flexible ways later.