frankenusb: improve flight control lock EICAS message
We now introduce the concept of a Captain's and FO's frankenusb
instance and use that to show whose flight control lock is enabled.

Ideally, we should have the tiller and lock notifications merged into
one message, but for now, we move the tiller (which will only be
active during taxi, so it can be more prominent) message to the
warning type.

Note to self: these EICAS messages have a maximum length. If you
exceed that when sending to PSX, the message will not be updated.