frankenrouter: re-enable PSX elevation database if no Qi198 injection
Problem: if all slave sims stop sending Qi198 elevation data to PSX,
PSX ground elevation will stay frozen at whatever level it was at the
time. This will result in a very large surprise when landing and the
ground is nowhere near where you expect it to be.

We already have a warning for this situation, but this commit adds
another safety net - if no Qi198 injection has been seen in 60s, we
send Qi198=-9999999, which tells PSX to start using its own elevation
database.

The PSX database will not match MSFS exactly, but it will at least not
be hundreds or thousands of feet off like it can be when the elevation
is frozen at some random value.