frankenrouter: change name= handling
After some discussions with Gary, we decided that proper addons send
"name=ID:display name" where ID is a short identifier that's unique in
the sim if you have several addons of the same type. Display name is a
more human-readable name. Both may be displayed by e.g a router
dashboard, so we expose both in the API.

We keep the display_name for now (which is what the router text status
display will use, and that one can also be set to "auth required" and
other things that are not extracted from the name= message.