frankenrouter: improve client welcome
These changes are based on a comparison of the data sent to a new
client by the PSX main server and by the router.

In order to avoid subtle bugs, we want to make sure the router sends
data that looks (and is ordered) identical to what a PSX main server
sends.

- queue any messages that arrive while the welcome is being sent
- use routercache timestamps to decide if START variables have arrived
- send queued messages once the welcome has been sent