frankenrouter: fix situ load bug
When a new client connects, we need to send "start" upstream to get
the latest values of the START type variables, and then send this to
the new client.

But we cannot send these variables to other clients as that would
confuse them.

So we handle this by always filter START variables but explicitly send
them during the client welcome.

However, START variables must be sent to all clients during a situ
load, otherwise the internal PSX state is not set correctly, giving us
all kinds of strange problems (e.g wrong N1, cabin pressure, hydraulic
faults, ...)

This commit ensures that we detect that we're in a situ load (by load1
seen more recently than load3) and then disable START variable
filtering.