frankenrouter: add a pause_forwarding flag to clients
We will use this instad of the welcome_sent message to signal that we
should queue incoming messages that would normally be forwarded to the
client.

We need to hold such messages while we welcome the client, but also
while we are sending a bang reply.

It seems that forwarding updates while the bang reply was being sent
was the trigger for the situ save problems.

Also stop incorrectly mark start variables as sent.